backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==3.2.0
passlib[bcrypt]
python-multipart==0.0.9
razorpay==1.4.1
python-dotenv==1.0.1
alembic==1.13.1
pydantic-settings==2.2.1
httpx==0.27.0